+++ app/components/FrameComponent/index.jsx
@@ -236,11 +236,19 @@ export function FrameComponent ({ widgetId, uri, widgets, data, onUpdate, enable
       return
     }
 
-    contentWindow.addEventListener('keydown', shortcuts.registerKeyDown)
-    contentWindow.addEventListener('keyup', shortcuts.registerKeyUp)
+    function onKeyDown (e) {
+      shortcuts.registerKeyDown(e)
+    }
+
+    function onKeyUp (e) {
+      shortcuts.registerKeyUp(e)
+    }
+
+    contentWindow.addEventListener('keydown', onKeyDown)
+    contentWindow.addEventListener('keyup', onKeyUp)
     return () => {
-      contentWindow.removeEventListener('keydown', shortcuts.registerKeyDown)
-      contentWindow.removeEventListener('keyup', shortcuts.registerKeyUp)
+      contentWindow.removeEventListener('keydown', onKeyDown)
+      contentWindow.removeEventListener('keyup', onKeyUp)
     }
   }, [frameRef.current?.contentWindow])