+++ app/components/Header/index.jsx
@@ -11,6 +11,7 @@ import { Preferences } from '../Preferences'
 import { Icon } from '../Icon'
 
 import * as api from '../../api'
+import * as modalUtils from '../../utils/modals'
 import * as windowUtils from '../../utils/window'
 
 import './style.css'
@@ -37,13 +38,27 @@ export function Header ({ title = DEFAULT_TITLE, features }) {
   to open the palette
   */
   React.useEffect(() => {
+    function hasFeature (feature) {
+      if (!Array.isArray(features)) {
+        return false
+      }
+      return features.includes(feature)
+    }
+
     function onShortcut (shortcut) {
       switch (shortcut) {
         case 'openPalette':
-          setPaletteIsOpen(true)
+          if (hasFeature('palette')) {
+            setPaletteIsOpen(true)
+          }
           break
-        case 'openSettings':
-          setPrefsOpen(true)
+        case 'openPreferences':
+          if (modalUtils.hasOpenModal()) {
+            return
+          }
+          if (hasFeature('preferences')) {
+            setPrefsOpen(true)
+          }
           break
       }
     }
@@ -61,7 +76,7 @@ export function Header ({ title = DEFAULT_TITLE, features }) {
       }
       teardown()
     }
-  }, [])
+  }, [features.join(',')])
 
   /**
    * Close the palette