+++ CHANGELOG.md
@@ -3,11 +3,20 @@
 ## 1.0.0-beta.10 [UNRELEASED]
 ### Added
 - Human/AI readable descriptors for types
-- Free run for LTC inputs
+- Free wheel for LTC inputs
 - Reference items now show a warning if they're missing a target id
+- Keyboard triggers
 ### Fixed
 - An issue where invalid types could crash the rundown widget
 - An issue where 0 (or falsy values) could not be used as ids for options in preferences of the select type
+- Preferences can no longer be opened in floating widget windows
+- Confirm dialogs now abort on escape
+- The shortcut for opening settings now works as intented
+- Settings no longer collide with other modals
+- An issue where children weren't removed when their parent group was removed
+### Changed
+- Keyboard shortcuts now use hotkeys-js for better stability
+- The item.apply event has been changed to item.change
 
 ## 1.0.0-beta.9
 ### Added